+++ src/library_browser.js
@@ -1214,10 +1214,7 @@ var LibraryBrowser = {
       _emscripten_get_window_title.buffer = _malloc(buflen);
     }
 
-    writeAsciiToMemory(
-      document.title.slice(0, buflen - 1),
-      _emscripten_get_window_title.buffer
-    );
+    stringToUTF8(document.title, _emscripten_get_window_title.buffer, buflen);
 
     return _emscripten_get_window_title.buffer;
   },
@@ -1225,7 +1222,7 @@ var LibraryBrowser = {
   emscripten_set_window_title__proxy: 'sync',
   emscripten_set_window_title__sig: 'vi',
   emscripten_set_window_title: function(title) {
-    setWindowTitle(AsciiToString(title));
+    setWindowTitle(UTF8ToString(title));
   },
 
   emscripten_get_screen_size__proxy: 'sync',