+++ ChangeLog.md
@@ -18,8 +18,17 @@ to browse the changes between the tags.
 
 See docs/process.md for more on how version tagging works.
 
-3.1.5
+3.1.6
 -----
+- Remove support for deprecated `EMMAKEN_COMPILER`, `EMMAKEN_CFLAGS`, and
+  `EMMAKEN_NO_SDK`  environment variables.  These are all legacy and redundant
+  in the face of other settings/flags:
+   - `EMMAKEN_COMPILER` -> `LLVM_ROOT` in the config settings
+   - `EMMAKEN_CFLAGS` -> `EMCC_CFLAGS`
+   - `EMMAKEN_NO_SDK` -> standard `-nostdlib` and `-nostdinc` flags
+
+3.1.5 - 02/17/2022
+------------------
 - Emscripten no longer uses the `allocate()` runtime function.  For backwards
   compatabiliy with external JS code we still include this function by default
   but it will no longer be included in `-sSTRICT` mode.  Usages of this function